Incorporate the recent changes to the generator
Add in the recent changes that were made to the code that generates ASCII art. These were made to the main project, but not added to the Pythonista app yet. I also took out the tests that ran when this script was run. The main code for image2ASCII includes those tests still, but they're not necessary for this version.